Рефакторит типы
