+++ project/src/types/state.ts
@@ -1,7 +1,7 @@
 import {AuthorizationStatus, CityName} from '../const';
-import type Offer from './offer';
+import type {Offer} from './offer';
 import type {Cities} from './city';
-import type AuthInfo from './auth-info';
+import type {AuthInfo} from './auth-info';
 
 type State = {
   activeCity: CityName;
@@ -12,4 +12,6 @@ type State = {
   isDataLoaded: boolean;
 }
 
-export default State;
+export type {
+  State
+};
